# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is the REST API to the PANDA database. \\n This is the only place to access data from the PANDA database.",
        "title": "PANDA REST API - localhost",
        "contact": {
            "name": "Jiří Švácha",
            "email": "jiri.svacha@eli-beams.eu"
        },
        "version": "1.0"
    },
    "host": "localhost:50000",
    "basePath": "/",
    "paths": {
...
        "/v1/publication": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create publication",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Publications"
                ],
                "summary": "Create publication",
                "parameters": [
                    {
                        "description": "Publication",
                        "name": "publication",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Publication"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Publication"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/v1/publication/wos/{doi}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get WOS data by DOI",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Publications"
                ],
                "summary": "Get WOS data by DOI",
                "parameters": [
                    {
                        "type": "string",
                        "description": "doi",
                        "name": "doi",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.WosAPIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/v1/publication/{uid}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get publication by uid",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Publications"
                ],
                "summary": "Get publication by uid",
                "parameters": [
                    {
                        "type": "string",
                        "description": "uid",
                        "name": "uid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Publication"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update publication",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Publications"
                ],
                "summary": "Update publication",
                "parameters": [
                    {
                        "type": "string",
                        "description": "uid",
                        "name": "uid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Publication",
                        "name": "publication",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Publication"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Publication"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete publication by uid",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Publications"
                ],
                "summary": "Delete publication by uid",
                "parameters": [
                    {
                        "type": "string",
                        "description": "uid",
                        "name": "uid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/v1/publications": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get publications",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Publications"
                ],
                "summary": "Get publications",
                "parameters": [
                    {
                        "type": "string",
                        "description": "search",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "pagination",
                        "name": "pagination",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Publication"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/v1/publications/export": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "CSV header: Media Type,Code,Experimental System,User Call,User Experiment,DOI,Web Link,Open Access Type,Title,Authors,Authors Count,ELI Authors,ELI Authors Count,Journal Title,Volume,Issue,Pages,Pages Count,Cite As,Impact Factor,Quartile Basis,Quartile,Year Of Publication,Date Of Publication,Abstract,Keywords,OECD Ford,Grant,WOS Number,ISSN,E-ISSN,EID Scopus,Publishing Country,Language,Note,UID",
                "produces": [
                    "text/csv"
                ],
                "tags": [
                    "Publications"
                ],
                "summary": "Get publications as CSV",
                "parameters": [
                    {
                        "type": "string",
                        "description": "search",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "CSV file"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/v1/researcher": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create researcher",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Researchers"
                ],
                "summary": "Create researcher",
                "parameters": [
                    {
                        "description": "Researcher",
                        "name": "researcher",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Researcher"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Researcher"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/v1/researcher/{uid}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get researcher by uid",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Researchers"
                ],
                "summary": "Get researcher by uid",
                "parameters": [
                    {
                        "type": "string",
                        "description": "uid",
                        "name": "uid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Researcher"
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update researcher",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Researchers"
                ],
                "summary": "Update researcher",
                "parameters": [
                    {
                        "type": "string",
                        "description": "uid",
                        "name": "uid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Researcher",
                        "name": "researcher",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Researcher"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Researcher"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete researcher by uid",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Researchers"
                ],
                "summary": "Delete researcher by uid",
                "parameters": [
                    {
                        "type": "string",
                        "description": "uid",
                        "name": "uid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.Researcher": {
            "type": "object",
            "properties": {
                "citizenship": {
                    "description": "citizenship is the country of citizenship of the researcher",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.Codebook"
                        }
                    ]
                },
                "firstName": {
                    "description": "firstName is the first name of the researcher",
                    "type": "string"
                },
                "identificationNumber": {
                    "description": "identificationNumber is the identification number of the researcher",
                    "type": "string"
                },
                "lastName": {
                    "description": "lastName is the last name of the researcher",
                    "type": "string"
                },
                "orcid": {
                    "description": "orcid is the ORCID identifier of the researcher",
                    "type": "string"
                },
                "researcherId": {
                    "description": "researcherId is the ResearcherID of the researcher",
                    "type": "string"
                },
                "scopusId": {
                    "description": "scopusId is the Scopus identifier of the researcher",
                    "type": "string"
                },
                "uid": {
                    "description": "uid is the unique identifier of the researcher",
                    "type": "string"
                },
                "updatedAt": {
                    "description": "updatedAt is the time when the researcher was last updated",
                    "type": "string"
                }
            }
        },
...
        "models.WosAPIResponse": {
            "type": "object",
            "properties": {
                "hits": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.WosHit"
                    }
                },
                "metadata": {
                    "$ref": "#/definitions/models.WosMetadata"
                }
            }
        },
...
        "models.WosMetadata": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT token. \u003cbr\u003e How to obtain: https://eli-eric.atlassian.net/wiki/spaces/CS/pages/948797504/How+to+get+PANDA+API+Token \u003cbr\u003e Add word Bearer before the token here.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}